2025/03/31
Added READ_FILE, KNR_STYLE, and FAR. Removed auto update library system because it will not work on older compilers. It's useless anyway. Updated CACHE or other library like removed #if or #undef like macros to make them support really old compilers. TODO: Improve ASM, for that I also need to create a cache library to detect which CPU, architecture and CPU version the user is using to use a single command. So, more work for me but later.