2025/12/25
Added CLEAR_INSTRUCTION_CACHE to FUNCTIONS. And added ROM to KEYWORDS. I didn't write it before I forgot but at the past push, I needed to re-name the project LIBCMT to CMT because it conflicts with Microsoft libcmt.lib. There is also something called CMT in Microsoft's stuff but it seems like not related to anything about programing so it is good. I fixed a lot of bugs, and documentation mistakes. I know I have to re-do the ASM. Right now I am creating FUNCTIONS that I'll use for OBJECTS including OBJECT.H. I am trying my best to complete this library. But I am not sure if I am wasting time. If you're reading this text, as you can see I am a bit in a situation I really wanna continue developing this library and also not sure how this will be seen or even if it will seen in the future... I just, don't know.